feat: #162 - filter for kids first doses
Done:
- update bulletin board
- update beans to deal with pecific tags
- upate GUI to isplay filter button with action sheet
- add items for accessibility purposes
- improve management of segmented controls

Note:
But still need refactoring about date because "prochain_rdv" field is for all doses an is always isplayed for each centre, a bit tricky to get the correct date back.
Maybe we should release this version andd then make a long / heavy refactoring so as to display the next date for the kids doses if it's relevant.

Signed-off-by: Pierre-Yves Lapersonne <dev@pylapersonne.info>